requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pyttsx3>=2.90
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
SpeechRecognition>=3.8.1
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-google-genai>=1.0.0
